dataflow: LiveVarAnalysis: Use funcdb and specially handle "main" function.
If returns information for function is available in funcdb, use that as a
true value of live out range for function. And for "main" function, use
a special condition that its live-out is empty (name(s) of such functions
of cause need to be configurable, not hardcoded as it is now).