dataflow: LiveVarAnalysis: Allow to perform underestimated analysis.
This kind of reinstates option to do it how it was done before, for
estmate_params pass, bit with extra clarifications to make sure that
callees don't affect use information. The idea of such pass is to
compure underestimated, but true set of live-in vars, i.e. registers
which are 100% parameters. This may be useful if for nothing else,
tehn to annotate results of proper, overestimating pass, which with
the lack of many callers, ends up with a big set of registers, some
of which are implausible params.