arch: ret_uses: Remove querying deprecated funcdb prop.
The code queries "ret" property which isn't really used in funcdb.
Instead, this functionality is now handled in dataflow.py, by using
"returns" property and setting it on CFG single exit.

Theoretically, setting uses for return statement in such a way may
be useful for playing with simplified passes instead of using the
full dataflow processing, and can be reintroduced later. Practically,
it's not useful given the more advanced processing handles that
feature now.

This change also means that arch.py doesn't depend on progdb.