core: Implement uses() and defs() on EXPR and COND.
This is allows to apply those functions polymorphically to all of EXPR,
Inst and BBlock (and RecursiveBlock with recent changes).

COND just defers to EXPR. And for EXPR, we assume it's side-effect free,
so defs() are always empty. This won't be true any longer when we start
to rewrite PseudoC call statements into C function call expressions.