graph, etc: Rename "dfsno" node property to "postno".
What is usually called "depth-first search number" is actually a preorder
number. But term "DFS number" is ambiguous and thus confusing, because
postorder is also pretty much a DFS number. Actually, taking it literally
word by word, "depth-first number" favors exactly postorder numbering.

This way or that, but what was called in ScratchABlock "dfsno" was actually
a postorder number, whereas in most of the litereature, preorder number is
meant. Solve this by renaming "dfsno" to "postno" and by avoiding using of
the term "DFS number" in the future. When needed, "preno" property will be
introduced.