sabl_cfg2pseudoc: New tool to convert a CFG graph to corresponding PseudoC.
This tool takes a (simplified) .dot file representing a CFG, and generates
a skeleton PseudoC program corresponding to that CFG. It uses generalized
"if" statement for >2 successors of a node and should generate correct
output. It doesn't depend on any 3rd-party lib, and does adhoc .dot parsing,
and thus support just a subset of line-oriented .dot format.