parser: Handle empty basic blocks, consisting of just a label.
E.g.:

label1:
label2:
label3:
    $a = 1

This will lead to creation of 3 basic blocks, linked sequentially:

label1 -> label2 -> label3

This (creation of empty basic blocks) may be not ideal, another option might
be to just treat label2 and label3 as aliases for label1. But it's definitely
better than to create empty basic blocks, but not link them to the successor,
as was done before.